+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Version 0.3.0
 - Fix DECIMAL automatic cast to float
 - Fix missing CastToTextNode
 - Fix missing cast from String in string cast nodes
+- Fix DECIMAL format text
 
 
 Version 0.2.0